requirements.txt
airflow-provider-great-expectations==0.2.7
apache-airflow-providers-apache-spark==4.3.0
great-expectations==0.17.23
numpy==1.26.1
openpyxl==3.1.2
pandas==2.1.2
pyspark==3.2.1
pytest
pytest-cov
python-telegram-bot==13.14
